Fix another part of tests after HandState refactor
