+++ server/src/test/kotlin/core/bettinground/FoldTest.kt
@@ -2,6 +2,7 @@ package core.bettinground
 
 import core.gameflow.*
 import core.gameflow.handstate.HandState
+import core.gameflow.handstate.toBuilder
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
